+++ modules/parallel/src/include/FutureObjectHelpers.hpp
@@ -11,15 +11,22 @@
 //=============================================================================
 #include <vector>
 #include "nlsParallel_exports.h"
-#include "Evaluator.hpp"
+#include "ArrayOf.hpp"
 #include "FutureObject.hpp"
+#include "Exception.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
 NLSPARALLEL_IMPEXP
-ArrayOfVector
-FutureFetchOutputs(
-    Evaluator* eval, const std::vector<FutureObject*>& futures, bool uniformOutput = false);
+ArrayOf
+FuturesToArrayOf(const std::vector<FutureObject*>& futures);
 //=============================================================================
+NLSPARALLEL_IMPEXP
+std::vector<FutureObject*>
+ArrayOfToFutures(const ArrayOf& _param);
+//=============================================================================
+NLSPARALLEL_IMPEXP
+ArrayOfVector
+vertCatArrayOfVector(const ArrayOfVector& args1, const ArrayOfVector& args2, Exception& e);
 }
 //=============================================================================